Factor out hasNext decoration from Iterable assertions
